2023.07.03 : 게시글 페이징 수정, 댓글 삭제 클릭시 취소 버튼 추가 해야할 것 : 1. 댓글 더보기에 있는 이전 댓글 수정, 삭제 수정 2. 관리자는 게시글, 댓글 수정 삭제 다 돼야함

+++ src/main/resources/static/js/ShareReplyMore.js
@@ -67,23 +67,27 @@ $(document).on('click', '.replyDelete', function() {
     }
 
     // 댓글 삭제 요청 보내기
-    $.ajax({
-        url: "/Music/Share/" + boardShareId + "/Reply/" + replyId,
-        type: "DELETE",
-        // CSRF 변수 값 가져오기
-        beforeSend: function(xhr) {
-            xhr.setRequestHeader(csrfHeader, csrfToken); // 토큰 헤더에 같이 보내는 것
-        },
-        success: function() {
-            alert("댓글을 삭제 하시겠습니까 ?");
-            // 페이지 새로고침
-            location.reload();
-        },
-        error: function() {
-            alert("댓글 삭제에 실패했습니다.");
-            error
-        }
-    });
+    if (confirm("댓글을 삭제 하시겠습니까?")) {
+        $.ajax({
+            url: "/Music/Share/" + boardShareId + "/Reply/" + replyId,
+            type: "DELETE",
+            // CSRF 변수 값 가져오기
+            beforeSend: function(xhr) {
+                xhr.setRequestHeader(csrfHeader, csrfToken); // 토큰 헤더에 같이 보내는 것
+            },
+            success: function() {
+                alert("댓글이 삭제되었습니다.");
+                // 페이지 새로고침
+                location.reload();
+            },
+            error: function() {
+                alert("댓글 삭제에 실패했습니다.");
+            }
+        });
+    } else {
+        // 삭제 취소
+        return false;
+    }
 });
 
 // 댓글 더보기
